+++ Changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## v1.11.0
+
+**New features:**
+- Added low-pass and high-pass filter coefficients calculation to `IIRFilters.c` to simplify initialization
+
+**Improvements:**
+- Moved faster math functions macros to `basicMath.h`
+
 ## v1.10.1
 
 **Improvements:**